requirements.txt
streamlit
openai
sentence-transformers
faiss-cpu
tiktoken
pandas
PyMuPDF
docx2txt